Added CLOSE_WAIT check
This fixes an issue where if EthernetServer.begin is called when there are any sockets in the CLOSE_WAIT state then it doesn't initialise the first socket that could be used (i.e. it would skip the CLOSE_WAIT socket). Particularly dangerous if all the sockets are in CLOSE_WAIT.