Added small delay in stop() method
Previously the stop() would tend to always take at least 1ms to execute due to the fact that the status() method is called almost immediately after disconnect(_sock). I experimentally determined that the disconnect() method needs at least about 0.25ms to complete in the best case (client on the same subnet, connected by cheap switch) so added a small delay. The effect is that the stop() method may now complete as quickly as 0.25ms.

Possibly a better way would be to replace the delay(1) with delayMicroseconds(250) but I wasn't sure what detrimental effects this could have due to polling 4x faster.